feat:all Profile integration tests
