+++ src/test/java/com/ensak/connect/integration/profile/ProfileControllerIntegrationTest.java
@@ -164,136 +164,4 @@ class ProfileControllerIntegrationTest extends AuthenticatedBaseIntegrationTest 
         //assertEquals(response.getMessage(),"{address=address is required, phone=phone is required, city=city is required, fullName=fullname is required, title=title is required}");
     }
 
-    @Test
-    public void itShouldGetSkillList() throws Exception{
-        //setup
-        var dummyUser = this.authenticateAsUser();
-        profileService.addSkill(dummyUser.getId(),
-                SkillRequestDTO.builder()
-                        .name("React")
-                        .level(Level.EXPERT)
-                        .build());
-        List<Skill> expectedResponse = new ArrayList<>();
-        expectedResponse.add(
-                Skill.builder()
-                .name("React")
-                .level(Level.EXPERT)
-                .build());
-
-        //request
-        String responseJson = api.perform(
-                get("/api/v1/profile/skills")
-                        .accept(MediaType.APPLICATION_JSON)
-                )
-                .andExpect(status().isOk())
-                .andReturn()
-                .getResponse()
-                .getContentAsString();
-
-        //assertions
-        List<Skill> response = objectMapper.readValue(responseJson,new TypeReference<List<Skill>>(){});
-        assertEquals(response.get(0).getLevel(),expectedResponse.get(0).getLevel());
-        assertEquals(response.get(0).getName(),expectedResponse.get(0).getName());
-
-    }
-
-    @Test
-    public void itShouldAddSkill() throws Exception{
-        //setup
-        var dummyUser = this.authenticateAsUser();
-        Skill expectedResponse = Skill.builder()
-                .name("React")
-                .level(Level.EXPERT)
-                .build();
-        //request
-
-        SkillRequestDTO payload = SkillRequestDTO.builder()
-                .name("React")
-                .level(Level.EXPERT)
-                .build();
-        String JsonPayload = objectMapper.writeValueAsString(payload);
-
-        String responseJson = api.perform(
-                        post("/api/v1/profile/skills")
-                                .accept(MediaType.APPLICATION_JSON)
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content(JsonPayload)
-                )
-                .andExpect(status().isCreated())
-                .andReturn()
-                .getResponse()
-                .getContentAsString();
-
-        //assertions
-        Skill response = objectMapper.readValue(responseJson,Skill.class);
-        assertEquals(response.getLevel(),expectedResponse.getLevel());
-        assertEquals(response.getName(),expectedResponse.getName());
-
-    }
-
-    @Test
-    public void itShouldNotAddSkill() throws Exception{
-        //setup
-        var dummyUser = this.authenticateAsUser();
-        //request
-
-        SkillRequestDTO payload = SkillRequestDTO.builder()
-                .level(Level.EXPERT)
-                .build();
-        String JsonPayload = objectMapper.writeValueAsString(payload);
-
-        String responseJson = api.perform(
-                        post("/api/v1/profile/skills")
-                                .accept(MediaType.APPLICATION_JSON)
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content(JsonPayload)
-                )
-                .andExpect(status().isBadRequest())
-                .andReturn()
-                .getResponse()
-                .getContentAsString();
-
-    }
-
-    @Test
-    public void itShouldDeleteUserSkill() throws Exception{
-        //setup
-        var dummyUser = this.authenticateAsUser();
-        Skill skill = profileService.addSkill(dummyUser.getId(),
-                SkillRequestDTO.builder()
-                        .name("React")
-                        .level(Level.EXPERT)
-                        .build());
-
-        //request
-        api.perform(
-                delete("/api/v1/profile/skills/"+skill.getId())
-                )
-                .andExpect(status().isNoContent());
-        //assertions
-        //System.out.println("->>>>"+profileService.getSkills(dummyUser.getId()));
-        assertTrue(profileService.getSkills(dummyUser.getId()).isEmpty());
-    }
-
-    @Test
-    public void itShouldNotFindUserSkillToDelete() throws Exception{
-        //setup
-        var dummyUser = this.authenticateAsUser();
-        Skill skill = profileService.addSkill(dummyUser.getId(),
-                SkillRequestDTO.builder()
-                        .name("React")
-                        .level(Level.EXPERT)
-                        .build());
-
-        //request
-        api.perform(
-                        delete("/api/v1/profile/skills/777")
-                )
-                .andExpect(status().isForbidden());
-    }
-
-
-
-
-
 }
